An Implementation of Anthropometric Measurement Standardisation Tests

Ensuring the precision and accuracy of measurements is critical when collecting anthropometric data. Anthropometrists are usually tested for precision and accuracy of measurement through standardisation tests performed prior to anthropometric data collection. This package provides functions to calculate inter- and intra-observer technical error of measurement (TEM) to assess precision of measurements.

See also

Author

Maintainer: Ernest Guevarra ernest@guevarra.io (ORCID)