Dataset from a standardisation exercise done in Liberia in preparation for a coverage survey.

Dataset from a standardisation exercise done in Liberia in preparation for a coverage survey.

Usage

liberiaStdData

Format

A data frame with 7 columns and 744 rows:

| **Variable** | **Description** | | :— | :— | | *eid* | Enumerator ID | | *cid* | Child ID | | *round* | Standardisation test round | | *weight* | Weight measurement | | *height* | Height measurement | | *muac* | Mid-upper arm circumference measurement | | *oedema* | Bilateral pitting oedema |

Examples

liberiaStdData
#>     eid cid round weight height  muac oedema
#> 1    27   1     1  12.00   93.5  14.4      0
#> 2    27   2     1  10.90   86.9  13.9      0
#> 3    27   3     1   9.30   78.1  13.8      0
#> 4    27   4     1  13.50   89.5  16.0      0
#> 5    27   5     1  11.90   83.2  16.7      0
#> 6    27   6     1  13.40   97.0  16.8      0
#> 7    27   7     1  15.20  101.0  17.3      0
#> 8    27   8     1   8.40   70.8  14.7      0
#> 9    27   9     1  13.20   93.8  16.2      0
#> 10   27  10     1  13.90  100.3  15.4      0
#> 11   27  11     1     NA  105.5  16.3      0
#> 12   27  12     1     NA   94.4  17.4      0
#> 13   28   1     1  12.00   93.5  14.2      0
#> 14   28   2     1  10.90   86.8  13.8      0
#> 15   28   3     1   9.30   77.2  13.7      0
#> 16   28   4     1  13.50   89.4  15.9      0
#> 17   28   5     1  11.90   83.2  16.4      0
#> 18   28   6     1  13.40   91.7  16.5      0
#> 19   28   7     1  15.20  101.0  17.0      0
#> 20   28   8     1   8.40   71.2  14.5      0
#> 21   28   9     1  13.20   93.7  16.2      0
#> 22   28  10     1  13.80  100.0  15.1      0
#> 23   28  11     1     NA  105.6  16.3      0
#> 24   28  12     1     NA   94.5  17.2      0
#> 25    2   1     1  11.90   94.4  15.2      0
#> 26    2   2     1  10.70   85.5  14.4      0
#> 27    2   3     1   9.30   78.0  14.5      0
#> 28    2   4     1  13.60   90.0  16.0      0
#> 29    2   5     1  11.90   85.9  16.2      0
#> 30    2   6     1  12.60   92.0  16.4      0
#> 31    2   7     1  15.10  100.0  17.2      0
#> 32    2   8     1   8.40   73.0  14.2      0
#> 33    2   9     1  13.20   93.3  16.0      0
#> 34    2  10     1  14.00  100.0  15.1      0
#> 35    2  11     1     NA  105.4  16.4      0
#> 36    2  12     1     NA   95.0  17.0      0
#> 37    1   1     1  12.00   95.1  14.3      0
#> 38    1   2     1  10.70   85.4  14.2      0
#> 39    1   3     1   9.40   77.1  14.5      0
#> 40    1   4     1  13.60   89.5  16.2      0
#> 41    1   5     1  12.00   85.8  16.3      0
#> 42    1   6     1  13.40   93.2  16.3      0
#> 43    1   7     1  15.00  101.0  17.3      0
#> 44    1   8     1   8.40   73.2  14.2      0
#> 45    1   9     1  13.20   93.3  16.3      0
#> 46    1  10     1  14.00  100.2  15.2      0
#> 47    1  11     1     NA  105.0  16.4      0
#> 48    1  12     1     NA   95.2  17.2      0
#> 49   30   1     1  11.80   93.0  14.0      0
#> 50   30   2     1  10.90   88.1  15.1      0
#> 51   30   3     1   9.40   81.1  14.0      0
#> 52   30   4     1  13.50   89.2  16.0      0
#> 53   30   5     1  12.50   90.0  15.3      0
#> 54   30   6     1  13.30   91.1  18.3      0
#> 55   30   7     1  15.30  101.0  17.2      0
#> 56   30   8     1   8.40   71.3  15.3      0
#> 57   30   9     1  13.30   94.0  16.2      0
#> 58   30  10     1  13.90  100.0  15.5      0
#> 59   30  11     1     NA  104.0  18.5      0
#> 60   30  12     1     NA   94.0  17.3      0
#> 61   29   1     1  11.80   93.0  14.0      0
#> 62   29   2     1  10.90   88.0  15.0      0
#> 63   29   3     1   9.20   81.1  14.2      0
#> 64   29   4     1  13.50   89.0  16.0      0
#> 65   29   5     1  12.50   90.0  15.3      0
#> 66   29   6     1  13.40   91.0  18.3      0
#> 67   29   7     1  15.31  101.0  18.3      0
#> 68   29   8     1   8.40   71.2  15.3      0
#> 69   29   9     1  13.30   94.0  16.0      0
#> 70   29  10     1  13.90  100.0  15.5      0
#> 71   29  11     1     NA  104.0  18.4      0
#> 72   29  12     1     NA   94.3  17.3      0
#> 73   23   1     1  12.00   93.5  14.5      0
#> 74   23   2     1  10.80   87.5  14.5      0
#> 75   23   3     1   9.50   78.0  14.2      0
#> 76   23   4     1  13.50   91.0  16.2      0
#> 77   23   5     1  12.00   85.5  16.7      0
#> 78   23   6     1  13.40   90.2  16.8      0
#> 79   23   7     1  15.20  101.0  16.9      0
#> 80   23   8     1   8.40   71.6  15.1      0
#> 81   23   9     1  13.20   93.8  15.4      0
#> 82   23  10     1  14.00   99.5  15.0      0
#> 83   23  11     1     NA  102.5  16.8      0
#> 84   23  12     1     NA   94.0  17.7      0
#> 85   24   1     1  12.00   93.2  14.2      0
#> 86   24   2     1  10.80   87.3  14.6      0
#> 87   24   3     1   9.50   78.0  13.9      0
#> 88   24   4     1  13.50   90.7  16.1      0
#> 89   24   5     1  11.90   85.1  16.5      0
#> 90   24   6     1  13.40   92.3  16.6      0
#> 91   24   7     1  15.30  101.1  17.8      0
#> 92   24   8     1   8.40   69.9  14.1      0
#> 93   24   9     1  13.20   93.6  16.1      0
#> 94   24  10     1  13.90  100.8  15.1      0
#> 95   24  11     1     NA  105.0  16.7      0
#> 96   24  12     1     NA   94.1  17.2      0
#> 97   15   1     1  12.00   93.3  14.0      0
#> 98   15   2     1  10.90   88.3  14.1      0
#> 99   15   3     1   9.30   78.0  13.9      0
#> 100  15   4     1  13.50   89.2  16.2      0
#> 101  15   5     1  11.90   82.5  16.6      0
#> 102  15   6     1  13.40   91.1  16.6      0
#> 103  15   7     1  15.30  101.2  17.0      0
#> 104  15   8     1   8.40   71.2  14.0      0
#> 105  15   9     1  13.90   93.7  16.1      0
#> 106  15  10     1  13.90  100.2  14.6      0
#> 107  15  11     1     NA  104.7  16.6      0
#> 108  15  12     1     NA   94.2  17.0      0
#> 109  16   1     1  12.00   93.3  14.0      0
#> 110  16   2     1  10.90   88.5  14.1      0
#> 111  16   3     1   9.50   77.5  14.0      0
#> 112  16   4     1  13.50   89.2  16.2      0
#> 113  16   5     1  11.90   82.5  16.6      0
#> 114  16   6     1  13.40   91.2  16.4      0
#> 115  16   7     1  15.30  101.2  17.0      0
#> 116  16   8     1   8.40   72.1  14.2      0
#> 117  16   9     1  13.30   93.9  16.1      0
#> 118  16  10     1  13.90   99.6  14.6      0
#> 119  16  11     1     NA  104.7  16.7      0
#> 120  16  12     1     NA   94.3  17.0      0
#> 121  20   1     1  12.00   93.5  14.0      0
#> 122  20   2     1  10.80   84.5  13.6      0
#> 123  20   3     1   9.30   78.0  14.3      0
#> 124  20   4     1  13.50   89.4  16.0      0
#> 125  20   5     1  11.90   86.0  16.4      0
#> 126  20   6     1  13.40   92.0  16.5      0
#> 127  20   7     1  15.20  101.0  17.5      0
#> 128  20   8     1   8.40   71.0  14.3      0
#> 129  20   9     1  13.20   94.0  16.1      0
#> 130  20  10     1  13.90  100.3  15.0      0
#> 131  20  11     1     NA  104.2  16.5      0
#> 132  20  12     1     NA   94.0  17.0      0
#> 133  19   1     1  12.00   93.5  14.0      0
#> 134  19   2     1  10.80   84.5  13.6      0
#> 135  19   3     1   9.30   78.0  14.3      0
#> 136  19   4     1  13.50   89.4  16.0      0
#> 137  19   5     1  11.90   86.0  16.4      0
#> 138  19   6     1  13.40   92.0  16.5      0
#> 139  19   7     1  15.20  101.0  17.5      0
#> 140  19   8     1   8.40   71.0  14.3      0
#> 141  19   9     1  13.20   94.0  16.1      0
#> 142  19  10     1  13.90  100.3  15.0      0
#> 143  19  11     1     NA  104.2  16.5      0
#> 144  19  12     1     NA   94.0  17.0      0
#> 145   7   1     1  12.00   93.3  14.3      0
#> 146   7   2     1  10.90   87.0  14.5      0
#> 147   7   3     1   9.50   80.1  14.2      0
#> 148   7   4     1  13.70   89.5  16.2      0
#> 149   7   5     1  11.90   85.3  16.7      0
#> 150   7   6     1  13.40   91.4  17.3      0
#> 151   7   7     1  15.20  100.4  17.5      0
#> 152   7   8     1   8.40   70.0  14.5      0
#> 153   7   9     1  13.20   93.8  16.4      0
#> 154   7  10     1  13.90  100.1  15.4      0
#> 155   7  11     1     NA  105.3  13.6      0
#> 156   7  12     1     NA   94.2  17.4      0
#> 157   8   1     1  12.00   93.3  14.3      0
#> 158   8   2     1  10.90   98.8  14.6      0
#> 159   8   3     1   9.50   80.1  14.2      0
#> 160   8   4     1  13.70   89.5  16.2      0
#> 161   8   5     1  11.90   85.3  16.8      0
#> 162   8   6     1  13.40   91.4  16.6      0
#> 163   8   7     1  15.20  100.4  17.5      0
#> 164   8   8     1   8.40   71.0  14.6      0
#> 165   8   9     1  13.20   93.8  16.4      0
#> 166   8  10     1  13.90  100.1  15.4      0
#> 167   8  11     1     NA  105.2  16.7      0
#> 168   8  12     1     NA   94.1  17.4      0
#> 169  11   1     1  12.00   93.0  14.2      0
#> 170  11   2     1  10.80   89.0  14.2      0
#> 171  11   3     1   9.50    8.5  14.7      0
#> 172  11   4     1  13.40   16.2  16.5      0
#> 173  11   5     1  12.00   89.6  17.0      0
#> 174  11   6     1  13.40   91.2  17.8      0
#> 175  11   7     1  14.70  101.0  17.5      0
#> 176  11   8     1   8.40   72.0  14.4      0
#> 177  11   9     1  13.70   93.0  16.5      0
#> 178  11  10     1  14.10  100.0  15.4      0
#> 179  11  11     1     NA  105.0  16.7      0
#> 180  11  12     1     NA   94.4  18.5      0
#> 181  12   1     1  11.90   93.5  13.9      0
#> 182  12   2     1  10.90   89.0  14.0      0
#> 183  12   3     1   9.30   82.0  14.3      0
#> 184  12   4     1  13.40   89.6  16.2      0
#> 185  12   5     1  12.20   84.0  16.0      0
#> 186  12   6     1  13.40   92.1  16.6      0
#> 187  12   7     1  15.20  100.0  16.6      0
#> 188  12   8     1   8.40   72.2  14.0      0
#> 189  12   9     1  13.20   93.5  16.5      0
#> 190  12  10     1  14.10  100.0  15.0      0
#> 191  12  11     1     NA  101.1  16.5      0
#> 192  12  12     1     NA   94.3  17.0      0
#> 193   3   1     1  12.90   93.0  14.3      0
#> 194   3   2     1  10.80   88.0  14.2      0
#> 195   3   3     1  94.40   77.0  14.2      0
#> 196   3   4     1  13.50   90.0  16.6      0
#> 197   3   5     1  11.90   86.0  17.4      0
#> 198   3   6     1  13.40   92.0  17.3      0
#> 199   3   7     1  15.10  100.0  17.4      0
#> 200   3   8     1   8.40   71.0  15.4      0
#> 201   3   9     1  13.20   94.0  16.6      0
#> 202   3  10     1  13.90  100.0  15.3      0
#> 203   3  11     1     NA  105.0  17.5      0
#> 204   3  12     1     NA   94.0  16.4      0
#> 205  26   1     1  12.00   93.3  14.1      0
#> 206  26   2     1  10.90   86.5  14.1      0
#> 207  26   3     1   9.40   79.0  14.6      0
#> 208  26   4     1  13.60   89.5  16.2      0
#> 209  26   5     1  12.00   84.4  16.3      0
#> 210  26   6     1  13.30   91.8  16.7      0
#> 211  26   7     1  14.20  100.2  17.2      0
#> 212  26   8     1   8.30   71.2  14.7      0
#> 213  26   9     1  13.20   93.4  16.1      0
#> 214  26  10     1  13.90  100.7  15.4      0
#> 215  26  11     1     NA  104.7  16.9      0
#> 216  26  12     1     NA   94.4  17.7      0
#> 217   4   1     1  11.90   93.5  14.1      0
#> 218   4   2     1  10.80   86.9  13.8      0
#> 219   4   3     1   9.30   77.6  13.9      0
#> 220   4   4     1  13.50   86.9  16.1      0
#> 221   4   5     1  11.90   85.7  16.5      0
#> 222   4   6     1  13.40   91.8  16.9      0
#> 223   4   7     1  15.10  100.8  17.1      0
#> 224   4   8     1   8.40   70.8  14.6      0
#> 225   4   9     1  13.20   94.2  16.2      0
#> 226   4  10     1  13.90  100.5  15.5      0
#> 227   4  11     1     NA  105.6  17.0      0
#> 228   4  12     1     NA   94.3  17.2      0
#> 229  25   1     1  12.00   93.3  14.1      0
#> 230  25   2     1  10.80   86.5  14.1      0
#> 231  25   3     1   9.20   79.0  14.6      0
#> 232  25   4     1  13.60   89.5  16.2      0
#> 233  25   5     1  12.00   84.4  16.3      0
#> 234  25   6     1  13.30   91.8  16.7      0
#> 235  25   7     1  15.20  100.2  17.2      0
#> 236  25   8     1   8.30   71.2  14.7      0
#> 237  25   9     1  13.30   93.4  16.1      0
#> 238  25  10     1  13.90  100.7  15.4      0
#> 239  25  11     1     NA  104.7  16.9      0
#> 240  25  12     1     NA   94.4  17.7      0
#> 241   5   1     1  11.90   92.2  13.4      0
#> 242   5   2     1  10.90   86.5  14.4      0
#> 243   5   3     1   9.30   80.0  14.1      0
#> 244   5   4     1  13.60   88.5  16.3      0
#> 245   5   5     1  11.90   87.6  16.4      0
#> 246   5   6     1  13.60   90.5  16.5      0
#> 247   5   7     1  15.00   98.2  17.3      0
#> 248   5   8     1   8.40   71.2  15.2      0
#> 249   5   9     1  13.20   98.2  16.3      0
#> 250   5  10     1  14.00   98.5  14.4      0
#> 251   5  11     1     NA  104.0  16.5      0
#> 252   5  12     1     NA   93.2  17.3      0
#> 253   6   1     1  11.90   92.6  14.1      0
#> 254   6   2     1  10.80   86.8  14.2      0
#> 255   6   3     1   9.30   79.6  13.5      0
#> 256   6   4     1  13.70   89.2  15.7      0
#> 257   6   5     1  11.90   86.9  16.4      0
#> 258   6   6     1  13.30   91.1  16.5      0
#> 259   6   7     1  15.00   99.3  17.2      0
#> 260   6   8     1   8.40   71.3  14.0      0
#> 261   6   9     1  13.20   93.2  16.4      0
#> 262   6  10     1  14.10   99.9  15.0      0
#> 263   6  11     1     NA  104.5  16.6      0
#> 264   6  12     1     NA   93.9  17.0      0
#> 265  13   1     1  12.00   92.9  14.8      0
#> 266  13   2     1  10.90   85.5  15.2      0
#> 267  13   3     1   9.30   81.3  14.5      0
#> 268  13   4     1  13.60   91.0  16.2      0
#> 269  13   5     1   8.30   85.5  16.3      0
#> 270  13   6     1  13.40   91.2  17.4      0
#> 271  13   7     1  15.00  100.0  18.2      0
#> 272  13   8     1   8.40   72.1  13.3      0
#> 273  13   9     1  13.30   93.3  16.5      0
#> 274  13  10     1  13.90   99.2  14.2      0
#> 275  13  11     1     NA  104.3  16.2      0
#> 276  13  12     1     NA   93.3  17.4      0
#> 277  14   1     1  11.90   92.9  14.9      0
#> 278  14   2     1  10.90   86.0  15.5      0
#> 279  14   3     1   9.30   81.2  14.5      0
#> 280  14   4     1  13.60   91.0  16.2      0
#> 281  14   5     1   8.30   85.0  17.1      0
#> 282  14   6     1  13.40   91.0  18.1      0
#> 283  14   7     1  15.10  100.2  17.7      0
#> 284  14   8     1   8.30   70.5  15.9      0
#> 285  14   9     1  13.30   94.0  15.8      0
#> 286  14  10     1  12.20  100.5  15.8      0
#> 287  14  11     1     NA  105.0  17.0      0
#> 288  14  12     1     NA   94.0  18.5      0
#> 289  21   1     1  12.00   92.4  13.5      0
#> 290  21   2     1  10.80   86.4  13.6      0
#> 291  21   3     1   9.50   77.5  13.2      0
#> 292  21   4     1  13.70   88.1  15.7      0
#> 293  21   5     1  12.00   85.1  16.2      0
#> 294  21   6     1  13.80   91.5  16.2      0
#> 295  21   7     1  15.40  100.4  16.5      0
#> 296  21   8     1   8.40   71.4  14.5      0
#> 297  21   9     1  13.20   93.4  15.5      0
#> 298  21  10     1  13.90   99.5  14.9      0
#> 299  21  11     1     NA  105.2  16.4      0
#> 300  21  12     1     NA    2.9  16.7      0
#> 301  22   1     1  12.10   93.2  14.0      0
#> 302  22   2     1  10.90   86.7  14.3      0
#> 303  22   3     1   9.40   77.9  14.3      0
#> 304  22   4     1  13.50   89.4  16.3      0
#> 305  22   5     1  11.90   84.5  17.1      0
#> 306  22   6     1  13.60   92.4  17.5      0
#> 307  22   7     1  15.20  101.0  17.3      0
#> 308  22   8     1   8.40   72.3  14.2      0
#> 309  22   9     1  13.20   93.5  16.7      0
#> 310  22  10     1  13.90   99.4  15.5      0
#> 311  22  11     1     NA  105.5  16.6      0
#> 312  22  12     1     NA   94.2  17.0      0
#> 313  31   1     1  11.90   93.0  14.4      0
#> 314  31   2     1  10.80   89.0  14.7      0
#> 315  31   3     1   9.30   78.0  14.3      0
#> 316  31   4     1  13.70   89.7  16.4      0
#> 317  31   5     1  12.00   86.5  16.6      0
#> 318  31   6     1  13.40   92.0  17.3      0
#> 319  31   7     1  15.20  100.8  17.5      0
#> 320  31   8     1   8.40   71.1  15.5      0
#> 321  31   9     1  13.10   94.3  16.6      0
#> 322  31  10     1  13.90  100.5  15.5      0
#> 323  31  11     1     NA  105.4  17.5      0
#> 324  31  12     1     NA   94.4  17.1      0
#> 325   9   1     1  12.00   91.0  15.0      0
#> 326   9   2     1  10.80   88.0  15.4      0
#> 327   9   3     1   8.90   80.1  13.6      0
#> 328   9   4     1  13.50   89.7  16.2      0
#> 329   9   5     1  11.90   88.7  17.1      0
#> 330   9   6     1  13.40   94.7  17.0      0
#> 331   9   7     1  12.90  100.7  16.4      0
#> 332   9   8     1   8.30   73.0  14.7      0
#> 333   9   9     1  13.30   94.8  16.8      0
#> 334   9  10     1  13.90   87.3  16.4      0
#> 335   9  11     1     NA  100.6  17.2      0
#> 336   9  12     1     NA   90.2  17.4      0
#> 337  10   1     1  11.60   93.0  14.6      0
#> 338  10   2     1  11.10   86.0  14.3      0
#> 339  10   3     1   9.30   80.1  14.4      0
#> 340  10   4     1  13.10   89.0  16.0      0
#> 341  10   5     1  11.90   86.0  17.4      0
#> 342  10   6     1  13.90   94.0  17.4      0
#> 343  10   7     1  15.30  100.0  17.4      0
#> 344  10   8     1   8.30   73.0  15.1      0
#> 345  10   9     1  13.30   91.0  17.0      0
#> 346  10  10     1  14.10  101.0  16.5      0
#> 347  10  11     1     NA  106.0  18.1      0
#> 348  10  12     1     NA   95.0  17.0      0
#> 349  17   1     1  12.00   93.0  14.0      0
#> 350  17   2     1  11.00   86.2  15.0      0
#> 351  17   3     1   9.30   79.0  13.5      0
#> 352  17   4     1  13.50   90.0  16.0      0
#> 353  17   5     1  12.20   87.0  15.0      0
#> 354  17   6     1  13.30   91.0  16.0      0
#> 355  17   7     1  15.20  100.4  17.0      0
#> 356  17   8     1   8.40   70.0  14.4      0
#> 357  17   9     1  13.30   94.0  15.3      0
#> 358  17  10     1  13.90  100.0  15.0      0
#> 359  17  11     1     NA  104.0  16.0      0
#> 360  17  12     1     NA   94.0  16.4      0
#> 361  18   1     1  12.00   94.0  14.5      0
#> 362  18   2     1  10.90   85.0  14.0      0
#> 363  18   3     1   9.40   79.0  13.5      0
#> 364  18   4     1  13.50   90.0  15.3      0
#> 365  18   5     1  12.20   87.0  15.5      0
#> 366  18   6     1  13.50   91.0  16.2      0
#> 367  18   7     1  15.20  100.0  17.0      0
#> 368  18   8     1   8.50   70.0  14.0      0
#> 369  18   9     1  13.40   93.0  16.5      0
#> 370  18  10     1  13.90   99.0  15.0      0
#> 371  18  11     1     NA  103.0  16.0      0
#> 372  18  12     1     NA   94.0  16.5      0
#> 373  27   1     2  12.30   92.4  14.0     NA
#> 374  27   2     2  10.90   86.4  13.9     NA
#> 375  27   3     2   9.50   78.1  13.8     NA
#> 376  27   4     2  13.70   89.4  15.7     NA
#> 377  27   5     2  11.80   83.4  16.4     NA
#> 378  27   6     2  13.70   91.5  16.6     NA
#> 379  27   7     2  15.20  101.3  17.1     NA
#> 380  27   8     2   8.30   71.2  14.2     NA
#> 381  27   9     2  13.30   93.6  16.1     NA
#> 382  27  10     2  14.70  101.1  15.2     NA
#> 383  27  11     2     NA  104.7  17.1     NA
#> 384  27  12     2     NA   93.4  17.0     NA
#> 385  31   1     2  12.20   93.0  14.4     NA
#> 386  31   2     2  10.80   86.7  14.6     NA
#> 387  31   3     2   9.30   78.0  14.8     NA
#> 388  31   4     2  13.70   89.5  16.5     NA
#> 389  31   5     2  11.90   87.6  17.1     NA
#> 390  31   6     2  13.50   93.0  17.1     NA
#> 391  31   7     2  15.10  100.3  18.1     NA
#> 392  31   8     2   8.60   71.1  15.4     NA
#> 393  31   9     2  13.20   94.3  16.4     NA
#> 394  31  10     2  14.60  100.2  15.4     NA
#> 395  31  11     2     NA  105.5  17.7     NA
#> 396  31  12     2     NA   93.3  17.6     NA
#> 397  28   1     2  12.30   92.5  13.7     NA
#> 398  28   2     2  10.90   86.3  14.2     NA
#> 399  28   3     2   9.50   78.2  14.3     NA
#> 400  28   4     2  13.70   89.2  15.8     NA
#> 401  28   5     2  11.80   83.4  16.2     NA
#> 402  28   6     2  13.70   92.1  16.5     NA
#> 403  28   7     2  15.20  100.3  17.3     NA
#> 404  28   8     2   8.30   71.2  14.1     NA
#> 405  28   9     2  13.30   93.7  15.8     NA
#> 406  28  10     2  14.70  100.2  15.3     NA
#> 407  28  11     2     NA  105.1  16.5     NA
#> 408  28  12     2     NA   94.1  16.8     NA
#> 409   2   1     2  12.20   93.0  14.4     NA
#> 410   2   2     2  10.90   87.8  14.5     NA
#> 411   2   3     2   9.40   78.0  14.2     NA
#> 412   2   4     2  13.20   89.5  15.5     NA
#> 413   2   5     2  11.90   87.0  16.4     NA
#> 414   2   6     2  13.50   92.5  16.5     NA
#> 415   2   7     2  15.50  100.0  17.0     NA
#> 416   2   8     2   8.40   73.0  14.0     NA
#> 417   2   9     2  13.10   94.2  16.0     NA
#> 418   2  10     2  14.70   99.8  15.3     NA
#> 419   2  11     2     NA  105.0  16.2     NA
#> 420   2  12     2     NA   93.6  17.5     NA
#> 421   1   1     2  12.20   92.2  14.3     NA
#> 422   1   2     2  10.90   87.6  15.9     NA
#> 423   1   3     2   9.40   77.8  14.1     NA
#> 424   1   4     2  13.70   89.2  16.0     NA
#> 425   1   5     2  11.90   87.0  17.0     NA
#> 426   1   6     2  13.50   92.2  17.3     NA
#> 427   1   7     2  15.50  100.2  16.4     NA
#> 428   1   8     2   8.40   72.4  14.3     NA
#> 429   1   9     2  13.30   94.3  16.3     NA
#> 430   1  10     2  14.70   99.9  15.3     NA
#> 431   1  11     2     NA  105.1  17.8     NA
#> 432   1  12     2     NA   93.4  18.5     NA
#> 433  12   1     2  12.40   93.2  13.8     NA
#> 434  12   2     2  10.80   87.8  13.9     NA
#> 435  12   3     2   9.40   82.5  13.8     NA
#> 436  12   4     2  13.70   88.5  16.0     NA
#> 437  12   5     2  12.20   88.0  15.5     NA
#> 438  12   6     2  13.60   92.0  17.0     NA
#> 439  12   7     2  15.30  100.0  16.8     NA
#> 440  12   8     2   8.30   73.0  14.0     NA
#> 441  12   9     2  13.40   93.0  16.5     NA
#> 442  12  10     2  14.70  100.0  14.9     NA
#> 443  12  11     2     NA  105.0  16.5     NA
#> 444  12  12     2     NA   93.5  16.5     NA
#> 445  11   1     2  12.00   93.3  14.2     NA
#> 446  11   2     2  10.80   86.3  14.2     NA
#> 447  11   3     2   9.30   81.3  13.6     NA
#> 448  11   4     2  13.70   98.5  16.5     NA
#> 449  11   5     2  12.10   88.0  16.0     NA
#> 450  11   6     2  13.60   91.5  17.2     NA
#> 451  11   7     2  15.20  100.4  17.3     NA
#> 452  11   8     2   8.30   17.0  14.2     NA
#> 453  11   9     2  12.30   93.1  16.5     NA
#> 454  11  10     2  14.70  100.0  16.0     NA
#> 455  11  11     2     NA  105.4  16.5     NA
#> 456  11  12     2     NA   93.8  17.4     NA
#> 457  26   1     2  12.30   93.0  13.6     NA
#> 458  26   2     2  10.90   96.5  13.9     NA
#> 459  26   3     2   9.60   79.5  13.8     NA
#> 460  26   4     2  13.70   89.3  16.5     NA
#> 461  26   5     2  11.70   86.3  16.7     NA
#> 462  26   6     2  13.80   91.6  16.7     NA
#> 463  26   7     2  15.50  100.3  17.3     NA
#> 464  26   8     2   8.40   71.1  14.6     NA
#> 465  26   9     2  13.30   94.1  16.0     NA
#> 466  26  10     2  14.70   99.8  15.5     NA
#> 467  26  11     2     NA  104.4  16.8     NA
#> 468  26  12     2     NA   93.4  17.3     NA
#> 469  25   1     2  12.30   93.0  13.6     NA
#> 470  25   2     2  10.90   96.5  13.9     NA
#> 471  25   3     2   9.60   79.5  13.8     NA
#> 472  25   4     2  13.70   89.3  16.5     NA
#> 473  25   5     2  11.70   86.3  16.7     NA
#> 474  25   6     2  13.50   91.6  16.7     NA
#> 475  25   7     2  15.20  100.3  17.3     NA
#> 476  25   8     2   8.40   71.1  14.6     NA
#> 477  25   9     2  13.30   94.4  16.0     NA
#> 478  25  10     2  14.70   99.8  15.5     NA
#> 479  25  11     2     NA  104.4  16.8     NA
#> 480  25  12     2     NA   93.4  17.3     NA
#> 481  15   1     2  12.00   93.3  14.0     NA
#> 482  15   2     2  10.90   87.7  14.0     NA
#> 483  15   3     2   9.30   78.0  14.0     NA
#> 484  15   4     2  13.70   89.0  16.0     NA
#> 485  15   5     2  11.90   86.0  16.0     NA
#> 486  15   6     2  13.60   91.6  17.0     NA
#> 487  15   7     2  15.30  101.2  17.0     NA
#> 488  15   8     2   8.30   71.0  14.0     NA
#> 489  15   9     2  93.60   13.3  16.0     NA
#> 490  15  10     2  14.70  100.2  14.9     NA
#> 491  15  11     2     NA  104.6  17.0     NA
#> 492  15  12     2     NA   94.0  17.0     NA
#> 493  16   1     2  12.30   92.7  13.9     NA
#> 494  16   2     2  10.90   88.0  14.1     NA
#> 495  16   3     2   9.30   77.8  14.0     NA
#> 496  16   4     2  13.70   89.0  16.0     NA
#> 497  16   5     2  11.90   86.0  16.2     NA
#> 498  16   6     2  13.60   91.6  16.8     NA
#> 499  16   7     2  15.30  101.0  17.0     NA
#> 500  16   8     2   8.30   71.4  14.0     NA
#> 501  16   9     2  13.30   93.6  16.4     NA
#> 502  16  10     2  14.70  100.2  15.2     NA
#> 503  16  11     2     NA  104.6  17.0     NA
#> 504  16  12     2     NA   93.7  17.2     NA
#> 505  22   1     2  12.30   93.0  14.0     NA
#> 506  22   2     2  10.90   87.5  14.4     NA
#> 507  22   3     2   9.30   80.4  14.5     NA
#> 508  22   4     2  13.00   89.4  16.3     NA
#> 509  22   5     2  11.90   84.9  16.8     NA
#> 510  22   6     2  13.50   93.3  16.7     NA
#> 511  22   7     2  15.30  100.7  17.5     NA
#> 512  22   8     2   8.40   70.5  15.3     NA
#> 513  22   9     2  13.20   94.4  16.5     NA
#> 514  22  10     2  14.70  104.0  15.0     NA
#> 515  22  11     2     NA  104.3  17.0     NA
#> 516  22  12     2     NA   94.0  17.5     NA
#> 517  20   1     2  12.30   92.5  13.5     NA
#> 518  20   2     2  10.90   88.0  14.0     NA
#> 519  20   3     2   9.30   78.0  14.3     NA
#> 520  20   4     2  13.70   88.4  16.1     NA
#> 521  20   5     2  11.90   87.0  16.2     NA
#> 522  20   6     2  13.50   94.0  16.1     NA
#> 523  20   7     2  15.50  100.2  17.0     NA
#> 524  20   8     2   8.30   71.0  14.2     NA
#> 525  20   9     2  13.30   94.0  15.5     NA
#> 526  20  10     2  14.70  100.5  15.5     NA
#> 527  20  11     2     NA  105.0  16.5     NA
#> 528  20  12     2     NA   93.5 171.0     NA
#> 529  24   1     2  12.30   92.5  14.2     NA
#> 530  24   2     2  10.90   87.2  13.9     NA
#> 531  24   3     2   9.50   78.0  13.9     NA
#> 532  24   4     2  13.70   90.3  16.1     NA
#> 533  24   5     2  11.90   85.4  16.2     NA
#> 534  24   6     2  13.50   92.2  16.9     NA
#> 535  24   7     2  15.40  100.5  16.9     NA
#> 536  24   8     2   7.70   83.0  14.1     NA
#> 537  24   9     2  13.30   93.5  16.1     NA
#> 538  24  10     2  14.60  100.0  15.1     NA
#> 539  24  11     2     NA  102.8  16.6     NA
#> 540  24  12     2     NA   93.4  17.3     NA
#> 541  19   1     2  12.30   92.5  13.5     NA
#> 542  19   2     2  10.90   88.0  14.0     NA
#> 543  19   3     2   9.30   78.0  14.3     NA
#> 544  19   4     2  13.70   88.4  16.1     NA
#> 545  19   5     2  11.90   87.0  16.2     NA
#> 546  19   6     2  13.50   94.0  16.1     NA
#> 547  19   7     2  15.50  100.2  17.0     NA
#> 548  19   8     2   8.30   71.0  14.2     NA
#> 549  19   9     2  13.30   94.0  15.5     NA
#> 550  19  10     2  14.70  100.5  15.5     NA
#> 551  19  11     2     NA  105.0  16.5     NA
#> 552  19  12     2     NA   93.5  17.1     NA
#> 553  23   1     2  12.30   91.3  14.0     NA
#> 554  23   2     2  10.90   87.2  14.5     NA
#> 555  23   3     2   9.50   78.0  13.8     NA
#> 556  23   4     2  13.70   91.0  15.7     NA
#> 557  23   5     2  11.90   85.5  16.4     NA
#> 558  23   6     2  13.50   90.3  16.8     NA
#> 559  23   7     2  15.40  100.0  16.7     NA
#> 560  23   8     2   8.30   71.0  14.3     NA
#> 561  23   9     2  13.30   92.8  16.2     NA
#> 562  23  10     2  14.60   99.0  15.4     NA
#> 563  23  11     2     NA  102.5  16.8     NA
#> 564  23  12     2     NA   93.0  17.0     NA
#> 565   8   1     2  12.30   93.2  14.3     NA
#> 566   8   2     2  10.90   86.5  14.3     NA
#> 567   8   3     2   9.30   78.7  14.5     NA
#> 568   8   4     2  13.80   88.4  16.3     NA
#> 569   8   5     2  11.80   86.6  16.4     NA
#> 570   8   6     2  13.60   91.6  16.7     NA
#> 571   8   7     2  15.30  100.5  17.5     NA
#> 572   8   8     2   8.40   70.7  13.7     NA
#> 573   8   9     2  13.30   93.5  16.4     NA
#> 574   8  10     2  14.70   99.5  15.1     NA
#> 575   8  11     2     NA  104.5  17.0     NA
#> 576   8  12     2     NA   93.6  17.6     NA
#> 577   6   1     2  12.40   92.3  14.0     NA
#> 578   6   2     2  10.90   87.0  14.3     NA
#> 579   6   3     2   9.20   79.0  13.5     NA
#> 580   6   4     2  13.60   87.2  16.1     NA
#> 581   6   5     2  12.00   86.9  17.0     NA
#> 582   6   6     2  13.40   91.3  16.9     NA
#> 583   6   7     2  15.30   99.4  17.4     NA
#> 584   6   8     2   8.40   72.4  14.1     NA
#> 585   6   9     2  12.50   93.0  16.2     NA
#> 586   6  10     2  14.70   99.9  15.1     NA
#> 587   6  11     2     NA  103.9  16.5     NA
#> 588   6  12     2     NA   93.0  16.6     NA
#> 589   3   1     2  12.30   93.2  14.4     NA
#> 590   3   2     2  10.90   88.0  14.2     NA
#> 591   3   3     2   9.30   78.0  14.3     NA
#> 592   3   4     2  13.70   88.0  16.6     NA
#> 593   3   5     2  11.90   85.0  16.5     NA
#> 594   3   6     2  13.40   92.0  17.3     NA
#> 595   3   7     2  12.30  101.0  16.4     NA
#> 596   3   8     2   8.30   72.0  14.3     NA
#> 597   3   9     2  13.30   94.0  16.7     NA
#> 598   3  10     2  14.70  100.3  16.3     NA
#> 599   3  11     2     NA  105.0  17.5     NA
#> 600   3  12     2     NA   95.0  17.4     NA
#> 601   4   1     2  12.30   93.4  14.2     NA
#> 602   4   2     2  10.90   87.4  14.0     NA
#> 603   4   3     2   9.90   78.9  13.9     NA
#> 604   4   4     2  13.60   89.3  16.4     NA
#> 605   4   5     2  11.90   86.7  16.6     NA
#> 606   4   6     2  13.20   92.5  16.9     NA
#> 607   4   7     2  15.20  100.9  17.4     NA
#> 608   4   8     2   8.30   71.0  14.8     NA
#> 609   4   9     2  13.30   93.5  15.9     NA
#> 610   4  10     2  14.70  100.8  15.2     NA
#> 611   4  11     2     NA  105.4  16.9     NA
#> 612   4  12     2     NA   95.3  17.4     NA
#> 613   7   1     2  12.30   93.2  14.2     NA
#> 614   7   2     2  10.90   87.1  14.3     NA
#> 615   7   3     2   9.30   78.7  14.4     NA
#> 616   7   4     2  13.80   88.4  16.3     NA
#> 617   7   5     2  11.90   86.5  16.4     NA
#> 618   7   6     2  13.50   91.5  16.7     NA
#> 619   7   7     2  15.30  100.3  17.5     NA
#> 620   7   8     2   8.30   70.7  13.8     NA
#> 621   7   9     2  13.30   93.7  16.4     NA
#> 622   7  10     2  14.70   99.6  16.1     NA
#> 623   7  11     2     NA  104.3  17.0     NA
#> 624   7  12     2     NA   94.2  17.6     NA
#> 625  21   1     2  12.40   91.7  13.5     NA
#> 626  21   2     2  11.00   87.5  14.0     NA
#> 627  21   3     2   9.60   77.5  13.0     NA
#> 628  21   4     2  13.10   89.2  15.7     NA
#> 629  21   5     2  11.20   85.2  15.6     NA
#> 630  21   6     2  13.40   93.6  16.1     NA
#> 631  21   7     2  15.20  100.1  17.0     NA
#> 632  21   8     2   8.30   70.4  14.0     NA
#> 633  21   9     2  13.20   93.5  15.3     NA
#> 634  21  10     2  14.50   99.5  15.0     NA
#> 635  21  11     2     NA  104.4  16.0     NA
#> 636  21  12     2     NA   93.5  17.0     NA
#> 637   5   1     2  12.40   92.1  13.6     NA
#> 638   5   2     2  10.90   87.0  14.5     NA
#> 639   5   3     2   9.20   71.1  13.5     NA
#> 640   5   4     2  13.60   87.0  16.0     NA
#> 641   5   5     2  12.00   87.0  16.9     NA
#> 642   5   6     2  13.40   91.0  17.0     NA
#> 643   5   7     2  15.30   99.2  17.0     NA
#> 644   5   8     2   8.40   72.5  14.0     NA
#> 645   5   9     2  14.00   92.9  15.4     NA
#> 646   5  10     2  14.80   97.5  14.4     NA
#> 647   5  11     2     NA  103.0  16.4     NA
#> 648   5  12     2     NA   92.4  16.4     NA
#> 649  13   1     2  12.30   92.2  13.5     NA
#> 650  13   2     2  10.90   85.5  15.4     NA
#> 651  13   3     2   9.30   79.2  13.4     NA
#> 652  13   4     2  13.30   89.1  16.3     NA
#> 653  13   5     2  12.00   87.4  16.2     NA
#> 654  13   6     2  13.70   90.5  16.4     NA
#> 655  13   7     2  15.30   99.4  17.1     NA
#> 656  13   8     2   8.30   71.2  14.4     NA
#> 657  13   9     2  13.40   93.1  17.2     NA
#> 658  13  10     2  14.70  100.1  15.1     NA
#> 659  13  11     2     NA  104.2  17.3     NA
#> 660  13  12     2     NA   93.3  17.2     NA
#> 661  14   1     2  12.30   92.0  14.2     NA
#> 662  14   2     2  10.80   85.5  15.0     NA
#> 663  14   3     2   9.30   79.0  14.9     NA
#> 664  14   4     2  13.30   89.1  16.9     NA
#> 665  14   5     2  12.00   86.0  17.3     NA
#> 666  14   6     2  13.70   91.5  17.1     NA
#> 667  14   7     2  15.30  100.0  17.5     NA
#> 668  14   8     2   8.30   72.0  14.5     NA
#> 669  14   9     2  13.60   93.0  17.1     NA
#> 670  14  10     2  14.70  100.1  16.1     NA
#> 671  14  11     2     NA  104.0  17.0     NA
#> 672  14  12     2     NA   93.9  17.2     NA
#> 673  17   1     2  12.30   93.3  13.4     NA
#> 674  17   2     2  10.80   87.0  14.0     NA
#> 675  17   3     2   9.20   79.0  13.5     NA
#> 676  17   4     2  13.70   89.0  15.4     NA
#> 677  17   5     2  11.80   85.0  15.0     NA
#> 678  17   6     2  13.50   91.0  16.0     NA
#> 679  17   7     2  15.30  100.0  16.0     NA
#> 680  17   8     2   8.30   72.0  14.0     NA
#> 681  17   9     2  13.70   93.4  15.2     NA
#> 682  17  10     2  14.70   97.0  15.0     NA
#> 683  17  11     2     NA  105.0  17.0     NA
#> 684  17  12     2     NA   95.0  16.0     NA
#> 685  18   1     2  12.30   93.0  13.5     NA
#> 686  18   2     2  10.80   87.0  13.6     NA
#> 687  18   3     2   9.30   78.0  13.5     NA
#> 688  18   4     2  14.20   89.0  15.5     NA
#> 689  18   5     2  11.80   87.0  15.2     NA
#> 690  18   6     2  13.50   91.0  16.0     NA
#> 691  18   7     2  15.30  100.0  16.0     NA
#> 692  18   8     2   8.40   72.0  13.5     NA
#> 693  18   9     2  13.30   93.0  15.3     NA
#> 694  18  10     2  14.70   99.0  15.0     NA
#> 695  18  11     2     NA  103.0  16.5     NA
#> 696  18  12     2     NA   95.0  16.5     NA
#> 697  29   1     2  12.30   93.0  14.0     NA
#> 698  29   2     2  10.70   87.0  14.0     NA
#> 699  29   3     2   9.20   79.0  14.0     NA
#> 700  29   4     2  13.70   90.0  16.4     NA
#> 701  29   5     2  10.50   86.2  16.3     NA
#> 702  29   6     2  13.60   92.1  16.4     NA
#> 703  29   7     2  15.30  100.0  17.0     NA
#> 704  29   8     2   8.30   74.0  14.3     NA
#> 705  29   9     2  13.30   94.0  16.0     NA
#> 706  29  10     2  14.70  100.0  15.3     NA
#> 707  29  11     2     NA  106.0  16.5     NA
#> 708  29  12     2     NA   94.4  17.1     NA
#> 709  30   1     2  12.40   95.4  16.0     NA
#> 710  30   2     2  10.90   87.2  15.5     NA
#> 711  30   3     2   9.20   79.0  15.2     NA
#> 712  30   4     2  13.60   88.1  17.5     NA
#> 713  30   5     2  10.50   86.1  17.5     NA
#> 714  30   6     2  13.60   92.2  17.2     NA
#> 715  30   7     2  15.30  100.0  17.5     NA
#> 716  30   8     2   8.40   73.1  15.5     NA
#> 717  30   9     2  13.30   94.6  17.5     NA
#> 718  30  10     2  14.70   99.0  17.3     NA
#> 719  30  11     2     NA  104.0  18.0     NA
#> 720  30  12     2     NA   92.0  19.0     NA
#> 721  10   1     2  12.40   93.0  14.1     NA
#> 722  10   2     2  11.20   86.0  14.1     NA
#> 723  10   3     2   9.10   83.0  14.1     NA
#> 724  10   4     2  16.60   89.0  16.4     NA
#> 725  10   5     2  12.00   88.0  16.5     NA
#> 726  10   6     2  13.50   91.0  16.4     NA
#> 727  10   7     2  16.00  100.0  17.0     NA
#> 728  10   8     2   8.20   73.0  14.5     NA
#> 729  10   9     2  13.60   93.0  16.3     NA
#> 730  10  10     2  14.80  100.0  15.5     NA
#> 731  10  11     2     NA  105.0  17.1     NA
#> 732  10  12     2     NA   93.5  17.4     NA
#> 733   9   1     2  12.30   93.4  14.6     NA
#> 734   9   2     2  10.90   87.5  13.4     NA
#> 735   9   3     2   9.40   82.3  14.6     NA
#> 736   9   4     2  13.80   89.4  16.1     NA
#> 737   9   5     2  12.00   84.0  16.7     NA
#> 738   9   6     2  13.50   91.4  16.8     NA
#> 739   9   7     2  15.60  100.0  16.4     NA
#> 740   9   8     2   8.20   70.1  14.5     NA
#> 741   9   9     2  13.40   93.1  16.6     NA
#> 742   9  10     2  14.70   99.7  15.7     NA
#> 743   9  11     2     NA  105.8  16.3     NA
#> 744   9  12     2     NA   94.4  17.8     NA