Standardisation test data - sample 1

Example dataset from the Standardized Monitoring and Assessment of Relief and Transitions (SMART) capacity building toolbox found at <http://smartmethodology.org/survey-planning-tools/smart-capacity-building-toolbox/>. The dataset has been restructured into a less wide format compared to the original wide format.

Usage

smartStd

Format

A data frame with 8 columns and 110 rows:

| **Variable** | **Description** | | :— | :— | | *subject* | Unique identifier of subjects who were measured | | *observer* | Unique identifier for those who performed the measurements | | *weight1* | First weight measurement. Units in kilograms | | *weight2* | Second weight measurement. Units in kilograms | | *height1* | First height measurement. Units in centimetres | | *height2* | Second height measurement. Units in centimetres | | *muac1* | First mid-upper arm circumference measurement. Units in millimetres | | *muac* | Second mid-upper arm circumference measurement. Units in millimetres |

Source

Standardized Monitoring and Assessment of Relief and Transitions (SMART) capacity building toolbox found at <http://smartmethodology.org/survey-planning-tools/smart-capacity-building-toolbox/>.

Examples

smartStd
#>     subject observer weight1 weight2 height1 height2 muac1 muac2
#> 1         1        0    15.9    16.0   102.0   102.3   170   170
#> 2         2        0    14.2    14.3    94.8    94.3   160   158
#> 3         3        0    16.7    16.7    97.2    97.2   184   173
#> 4         4        0    14.1    14.2    93.1    93.2   174   164
#> 5         5        0    14.8    14.6    99.6    99.1   155   160
#> 6         6        0    16.2    16.1    98.9    98.3   173   170
#> 7         7        0    10.5    10.5    83.5    83.5   145   145
#> 8         8        0    16.5    16.4    98.0    98.0   170   171
#> 9         9        0    14.1    14.3    95.1    92.9   163   160
#> 10       10        0    13.5    13.5    94.7    94.2   162   160
#> 11        1        1    16.1    15.9   101.3   101.8   167   163
#> 12        2        1    14.1    14.2    94.4    94.3   152   154
#> 13        3        1    16.7    16.1    96.5    96.8   172   176
#> 14        4        1    13.7    13.4    92.4    93.0   154   154
#> 15        5        1    14.0    14.6    99.3    99.4   146   154
#> 16        6        1    16.0    15.8    98.1    19.2   168   169
#> 17        7        1    10.9    10.1    83.9    14.1   140   140
#> 18        8        1    16.3    15.5    98.0    97.8   168   169
#> 19        9        1    14.0    14.1    94.4    94.4   157   161
#> 20       10        1    13.2    13.5    94.4    93.9   154   154
#> 21        1        2    16.0    15.9   101.5   101.1   164   164
#> 22        2        2    14.2    14.2    94.6    93.6   145   143
#> 23        3        2    16.3    16.6    97.3    95.9   181   174
#> 24        4        2    14.1    14.0    93.2    92.8   157   151
#> 25        5        2    14.5    14.6   100.1    99.3   153   149
#> 26        6        2    15.9    16.2    97.8    98.2   164   163
#> 27        7        2    10.4    10.5    83.7   134.0   134   141
#> 28        8        2    16.3    16.2    97.6    97.2   165   165
#> 29        9        2    14.1    14.2    94.5    93.8   152   150
#> 30       10        2    13.4    13.6    93.9    94.2   152   152
#> 31        1        3    16.0    15.8   101.6   102.5   171   174
#> 32        2        3    14.3    14.2    94.2    94.3   153   155
#> 33        3        3    16.6    16.6    97.1    97.1   179   184
#> 34        4        3    14.2    14.1    92.6    93.1   155   154
#> 35        5        3    14.5    14.5   100.0   100.0   147    15
#> 36        6        3    16.0    16.1    98.9    98.6   172   170
#> 37        7        3    10.5    10.5    83.5    84.1   137   146
#> 38        8        3    16.3    16.3    98.1    98.2   170   170
#> 39        9        3    14.1    14.0    94.0    94.2   156   153
#> 40       10        3    13.5    13.5    94.8    95.2   152   154
#> 41        1        4    16.1    15.9   100.2   102.2   162   170
#> 42        2        4    14.3    13.9    94.3    94.8   157   150
#> 43        3        4    16.7    16.2    96.6    97.1   180   180
#> 44        4        4    14.2    14.1    92.4    93.1   151   157
#> 45        5        4    14.5    14.5    99.3    99.3   148   153
#> 46        6        4    16.0    15.9    98.4    98.4   168   167
#> 47        7        4    10.5    10.5    83.3    83.3   134   141
#> 48        8        4    16.4    16.1    97.4    97.8   170   171
#> 49        9        4    14.1    14.1    95.5    94.2   158   164
#> 50       10        4    13.4    13.5    95.3    93.7   158   155
#> 51        1        5    15.9    15.8   100.2   101.2   173   172
#> 52        2        5    14.4    14.2    92.5    94.1   156   157
#> 53        3        5    16.7    16.7    96.1    95.1   172   177
#> 54        4        5    14.1    14.1    92.2    92.8   155   155
#> 55        5        5    14.5    14.6    99.0    99.2   152   150
#> 56        6        5    10.0    16.2    97.9    97.2   169   167
#> 57        7        5    10.5    10.4    83.3    83.3   140   143
#> 58        8        5    16.3    16.3    97.2    97.6   168   173
#> 59        9        5    14.0    14.1    93.1    93.2   160   157
#> 60       10        5    13.5    13.4    93.5    93.7   157   196
#> 61        1        6    15.9    15.9   101.4   101.5   167   173
#> 62        2        6    14.4    14.4    93.2    93.1   156   151
#> 63        3        6    16.6    16.3    96.6    96.5   171   167
#> 64        4        6    14.3    14.2    92.2    91.7   152   155
#> 65        5        6    14.5    14.6    99.3    99.8   151   155
#> 66        6        6    16.1    16.2    98.4    97.7   168   175
#> 67        7        6    10.5    10.5    83.9    83.9   134   134
#> 68        8        6    16.8    16.3    97.2    97.1   171   176
#> 69        9        6    14.2    14.1    92.1    92.0   151   153
#> 70       10        6    13.6    13.6    93.6    94.1   151   152
#> 71        1        7    16.0    15.9   102.1   101.6   165   160
#> 72        2        7    14.2    14.1    93.7    94.2   148   155
#> 73        3        7    16.3    16.7    96.7    96.9   173   167
#> 74        4        7    14.2    14.3    92.5    93.0   159   151
#> 75        5        7    14.6    14.7    99.3    99.5   152   140
#> 76        6        7    16.1    16.1    98.6    98.0   169   164
#> 77        7        7    10.7    10.7    83.4    83.4   139   139
#> 78        8        7    16.3    16.4    98.5    97.7   170   175
#> 79        9        7    14.3    14.3    93.5    93.7   158   186
#> 80       10        7    13.7    13.7    94.0    93.5   156   152
#> 81        1        8    16.0    15.9   102.1   102.3   170   169
#> 82        2        8    14.3    14.2    94.4    94.3   150   152
#> 83        3        8    17.0    16.6    97.4    97.0   176   173
#> 84        4        8    14.2    14.2    93.1    92.5   151   159
#> 85        5        8    14.5    14.6    99.5   101.0   147   149
#> 86        6        8    16.2    16.1    98.8    98.4   170   165
#> 87        7        8    10.3    10.5    84.5    84.5   141   141
#> 88        8        8    16.4    16.3    98.0    98.0   171   172
#> 89        9        8    14.2    14.2    92.2    93.9   159   158
#> 90       10        8    13.6    13.7    94.0    94.4   166   156
#> 91        1        9    16.0    15.9   102.2   102.4   162   162
#> 92        2        9    14.2    14.2    93.9    93.7   154   159
#> 93        3        9    16.4    16.6    97.1    97.1   166   166
#> 94        4        9    14.2    14.2    92.9    92.6   148   148
#> 95        5        9    14.7    14.6    99.2    99.4   147   147
#> 96        6        9    16.1    16.1    98.6    98.5   162   162
#> 97        7        9    10.5    10.5    94.0    94.1   111   112
#> 98        8        9    16.4    16.3    97.8    97.2   167   176
#> 99        9        9    14.1    14.2    94.7    93.6   181   181
#> 100      10        9    14.1    13.7    94.5    93.6   150   150
#> 101       1       10    15.9    15.0   102.0   102.0   169   165
#> 102       2       10    14.2    14.3    94.1    93.9   154   152
#> 103       3       10    16.7    16.2    96.5    97.1   175   171
#> 104       4       10    14.2    14.1    93.0    93.0   155   153
#> 105       5       10    14.6    14.6    99.5    98.6   150   147
#> 106       6       10    16.1    16.1    97.5    97.8   170   165
#> 107       7       10    10.5    10.5    84.5    84.5   141   141
#> 108       8       10    16.4    16.3    97.5    97.5   169   171
#> 109       9       10    14.2    14.2    93.5    94.4   159   161
#> 110      10       10    14.5    13.7    94.7    94.6   155   156