Standardisation test data - sample 2

Example dataset from the Standardized Monitoring and Assessment of Relief and Transitions (SMART) capacity building toolbox found at <http://smartmethodology.org/survey-planning-tools/smart-capacity-building-toolbox/>. This dataset is the same as the smartStd dataset but has been restructured into a long format (as per tidyverse specifications) compared to original format.

Usage

smartStdLong

Format

A data frame with 5 columns and 660 rows:

subject

Subject unique identifier

observer

Observer unique identifier

measure_type

Type of measurement. Can be height, weight, or muac

measure_round

Measurement round. Can be 1 or 2 for first and second round of measurement respectively

measure_value

Measurement

Source

Standardized Monitoring and Assessment of Relief and Transitions (SMART) capacity building toolbox found at <http://smartmethodology.org/survey-planning-tools/smart-capacity-building-toolbox/>.

Examples

smartStdLong
#>     subject observer measure_type measure_round measure_value
#> 1         1        0       weight             1          15.9
#> 2         2        0       weight             1          14.2
#> 3         3        0       weight             1          16.7
#> 4         4        0       weight             1          14.1
#> 5         5        0       weight             1          14.8
#> 6         6        0       weight             1          16.2
#> 7         7        0       weight             1          10.5
#> 8         8        0       weight             1          16.5
#> 9         9        0       weight             1          14.1
#> 10       10        0       weight             1          13.5
#> 11        1        1       weight             1          16.1
#> 12        2        1       weight             1          14.1
#> 13        3        1       weight             1          16.7
#> 14        4        1       weight             1          13.7
#> 15        5        1       weight             1          14.0
#> 16        6        1       weight             1          16.0
#> 17        7        1       weight             1          10.9
#> 18        8        1       weight             1          16.3
#> 19        9        1       weight             1          14.0
#> 20       10        1       weight             1          13.2
#> 21        1        2       weight             1          16.0
#> 22        2        2       weight             1          14.2
#> 23        3        2       weight             1          16.3
#> 24        4        2       weight             1          14.1
#> 25        5        2       weight             1          14.5
#> 26        6        2       weight             1          15.9
#> 27        7        2       weight             1          10.4
#> 28        8        2       weight             1          16.3
#> 29        9        2       weight             1          14.1
#> 30       10        2       weight             1          13.4
#> 31        1        3       weight             1          16.0
#> 32        2        3       weight             1          14.3
#> 33        3        3       weight             1          16.6
#> 34        4        3       weight             1          14.2
#> 35        5        3       weight             1          14.5
#> 36        6        3       weight             1          16.0
#> 37        7        3       weight             1          10.5
#> 38        8        3       weight             1          16.3
#> 39        9        3       weight             1          14.1
#> 40       10        3       weight             1          13.5
#> 41        1        4       weight             1          16.1
#> 42        2        4       weight             1          14.3
#> 43        3        4       weight             1          16.7
#> 44        4        4       weight             1          14.2
#> 45        5        4       weight             1          14.5
#> 46        6        4       weight             1          16.0
#> 47        7        4       weight             1          10.5
#> 48        8        4       weight             1          16.4
#> 49        9        4       weight             1          14.1
#> 50       10        4       weight             1          13.4
#> 51        1        5       weight             1          15.9
#> 52        2        5       weight             1          14.4
#> 53        3        5       weight             1          16.7
#> 54        4        5       weight             1          14.1
#> 55        5        5       weight             1          14.5
#> 56        6        5       weight             1          10.0
#> 57        7        5       weight             1          10.5
#> 58        8        5       weight             1          16.3
#> 59        9        5       weight             1          14.0
#> 60       10        5       weight             1          13.5
#> 61        1        6       weight             1          15.9
#> 62        2        6       weight             1          14.4
#> 63        3        6       weight             1          16.6
#> 64        4        6       weight             1          14.3
#> 65        5        6       weight             1          14.5
#> 66        6        6       weight             1          16.1
#> 67        7        6       weight             1          10.5
#> 68        8        6       weight             1          16.8
#> 69        9        6       weight             1          14.2
#> 70       10        6       weight             1          13.6
#> 71        1        7       weight             1          16.0
#> 72        2        7       weight             1          14.2
#> 73        3        7       weight             1          16.3
#> 74        4        7       weight             1          14.2
#> 75        5        7       weight             1          14.6
#> 76        6        7       weight             1          16.1
#> 77        7        7       weight             1          10.7
#> 78        8        7       weight             1          16.3
#> 79        9        7       weight             1          14.3
#> 80       10        7       weight             1          13.7
#> 81        1        8       weight             1          16.0
#> 82        2        8       weight             1          14.3
#> 83        3        8       weight             1          17.0
#> 84        4        8       weight             1          14.2
#> 85        5        8       weight             1          14.5
#> 86        6        8       weight             1          16.2
#> 87        7        8       weight             1          10.3
#> 88        8        8       weight             1          16.4
#> 89        9        8       weight             1          14.2
#> 90       10        8       weight             1          13.6
#> 91        1        9       weight             1          16.0
#> 92        2        9       weight             1          14.2
#> 93        3        9       weight             1          16.4
#> 94        4        9       weight             1          14.2
#> 95        5        9       weight             1          14.7
#> 96        6        9       weight             1          16.1
#> 97        7        9       weight             1          10.5
#> 98        8        9       weight             1          16.4
#> 99        9        9       weight             1          14.1
#> 100      10        9       weight             1          14.1
#> 101       1       10       weight             1          15.9
#> 102       2       10       weight             1          14.2
#> 103       3       10       weight             1          16.7
#> 104       4       10       weight             1          14.2
#> 105       5       10       weight             1          14.6
#> 106       6       10       weight             1          16.1
#> 107       7       10       weight             1          10.5
#> 108       8       10       weight             1          16.4
#> 109       9       10       weight             1          14.2
#> 110      10       10       weight             1          14.5
#> 111       1        0       height             1         102.0
#> 112       2        0       height             1          94.8
#> 113       3        0       height             1          97.2
#> 114       4        0       height             1          93.1
#> 115       5        0       height             1          99.6
#> 116       6        0       height             1          98.9
#> 117       7        0       height             1          83.5
#> 118       8        0       height             1          98.0
#> 119       9        0       height             1          95.1
#> 120      10        0       height             1          94.7
#> 121       1        1       height             1         101.3
#> 122       2        1       height             1          94.4
#> 123       3        1       height             1          96.5
#> 124       4        1       height             1          92.4
#> 125       5        1       height             1          99.3
#> 126       6        1       height             1          98.1
#> 127       7        1       height             1          83.9
#> 128       8        1       height             1          98.0
#> 129       9        1       height             1          94.4
#> 130      10        1       height             1          94.4
#> 131       1        2       height             1         101.5
#> 132       2        2       height             1          94.6
#> 133       3        2       height             1          97.3
#> 134       4        2       height             1          93.2
#> 135       5        2       height             1         100.1
#> 136       6        2       height             1          97.8
#> 137       7        2       height             1          83.7
#> 138       8        2       height             1          97.6
#> 139       9        2       height             1          94.5
#> 140      10        2       height             1          93.9
#> 141       1        3       height             1         101.6
#> 142       2        3       height             1          94.2
#> 143       3        3       height             1          97.1
#> 144       4        3       height             1          92.6
#> 145       5        3       height             1         100.0
#> 146       6        3       height             1          98.9
#> 147       7        3       height             1          83.5
#> 148       8        3       height             1          98.1
#> 149       9        3       height             1          94.0
#> 150      10        3       height             1          94.8
#> 151       1        4       height             1         100.2
#> 152       2        4       height             1          94.3
#> 153       3        4       height             1          96.6
#> 154       4        4       height             1          92.4
#> 155       5        4       height             1          99.3
#> 156       6        4       height             1          98.4
#> 157       7        4       height             1          83.3
#> 158       8        4       height             1          97.4
#> 159       9        4       height             1          95.5
#> 160      10        4       height             1          95.3
#> 161       1        5       height             1         100.2
#> 162       2        5       height             1          92.5
#> 163       3        5       height             1          96.1
#> 164       4        5       height             1          92.2
#> 165       5        5       height             1          99.0
#> 166       6        5       height             1          97.9
#> 167       7        5       height             1          83.3
#> 168       8        5       height             1          97.2
#> 169       9        5       height             1          93.1
#> 170      10        5       height             1          93.5
#> 171       1        6       height             1         101.4
#> 172       2        6       height             1          93.2
#> 173       3        6       height             1          96.6
#> 174       4        6       height             1          92.2
#> 175       5        6       height             1          99.3
#> 176       6        6       height             1          98.4
#> 177       7        6       height             1          83.9
#> 178       8        6       height             1          97.2
#> 179       9        6       height             1          92.1
#> 180      10        6       height             1          93.6
#> 181       1        7       height             1         102.1
#> 182       2        7       height             1          93.7
#> 183       3        7       height             1          96.7
#> 184       4        7       height             1          92.5
#> 185       5        7       height             1          99.3
#> 186       6        7       height             1          98.6
#> 187       7        7       height             1          83.4
#> 188       8        7       height             1          98.5
#> 189       9        7       height             1          93.5
#> 190      10        7       height             1          94.0
#> 191       1        8       height             1         102.1
#> 192       2        8       height             1          94.4
#> 193       3        8       height             1          97.4
#> 194       4        8       height             1          93.1
#> 195       5        8       height             1          99.5
#> 196       6        8       height             1          98.8
#> 197       7        8       height             1          84.5
#> 198       8        8       height             1          98.0
#> 199       9        8       height             1          92.2
#> 200      10        8       height             1          94.0
#> 201       1        9       height             1         102.2
#> 202       2        9       height             1          93.9
#> 203       3        9       height             1          97.1
#> 204       4        9       height             1          92.9
#> 205       5        9       height             1          99.2
#> 206       6        9       height             1          98.6
#> 207       7        9       height             1          94.0
#> 208       8        9       height             1          97.8
#> 209       9        9       height             1          94.7
#> 210      10        9       height             1          94.5
#> 211       1       10       height             1         102.0
#> 212       2       10       height             1          94.1
#> 213       3       10       height             1          96.5
#> 214       4       10       height             1          93.0
#> 215       5       10       height             1          99.5
#> 216       6       10       height             1          97.5
#> 217       7       10       height             1          84.5
#> 218       8       10       height             1          97.5
#> 219       9       10       height             1          93.5
#> 220      10       10       height             1          94.7
#> 221       1        0         muac             1         170.0
#> 222       2        0         muac             1         160.0
#> 223       3        0         muac             1         184.0
#> 224       4        0         muac             1         174.0
#> 225       5        0         muac             1         155.0
#> 226       6        0         muac             1         173.0
#> 227       7        0         muac             1         145.0
#> 228       8        0         muac             1         170.0
#> 229       9        0         muac             1         163.0
#> 230      10        0         muac             1         162.0
#> 231       1        1         muac             1         167.0
#> 232       2        1         muac             1         152.0
#> 233       3        1         muac             1         172.0
#> 234       4        1         muac             1         154.0
#> 235       5        1         muac             1         146.0
#> 236       6        1         muac             1         168.0
#> 237       7        1         muac             1         140.0
#> 238       8        1         muac             1         168.0
#> 239       9        1         muac             1         157.0
#> 240      10        1         muac             1         154.0
#> 241       1        2         muac             1         164.0
#> 242       2        2         muac             1         145.0
#> 243       3        2         muac             1         181.0
#> 244       4        2         muac             1         157.0
#> 245       5        2         muac             1         153.0
#> 246       6        2         muac             1         164.0
#> 247       7        2         muac             1         134.0
#> 248       8        2         muac             1         165.0
#> 249       9        2         muac             1         152.0
#> 250      10        2         muac             1         152.0
#> 251       1        3         muac             1         171.0
#> 252       2        3         muac             1         153.0
#> 253       3        3         muac             1         179.0
#> 254       4        3         muac             1         155.0
#> 255       5        3         muac             1         147.0
#> 256       6        3         muac             1         172.0
#> 257       7        3         muac             1         137.0
#> 258       8        3         muac             1         170.0
#> 259       9        3         muac             1         156.0
#> 260      10        3         muac             1         152.0
#> 261       1        4         muac             1         162.0
#> 262       2        4         muac             1         157.0
#> 263       3        4         muac             1         180.0
#> 264       4        4         muac             1         151.0
#> 265       5        4         muac             1         148.0
#> 266       6        4         muac             1         168.0
#> 267       7        4         muac             1         134.0
#> 268       8        4         muac             1         170.0
#> 269       9        4         muac             1         158.0
#> 270      10        4         muac             1         158.0
#> 271       1        5         muac             1         173.0
#> 272       2        5         muac             1         156.0
#> 273       3        5         muac             1         172.0
#> 274       4        5         muac             1         155.0
#> 275       5        5         muac             1         152.0
#> 276       6        5         muac             1         169.0
#> 277       7        5         muac             1         140.0
#> 278       8        5         muac             1         168.0
#> 279       9        5         muac             1         160.0
#> 280      10        5         muac             1         157.0
#> 281       1        6         muac             1         167.0
#> 282       2        6         muac             1         156.0
#> 283       3        6         muac             1         171.0
#> 284       4        6         muac             1         152.0
#> 285       5        6         muac             1         151.0
#> 286       6        6         muac             1         168.0
#> 287       7        6         muac             1         134.0
#> 288       8        6         muac             1         171.0
#> 289       9        6         muac             1         151.0
#> 290      10        6         muac             1         151.0
#> 291       1        7         muac             1         165.0
#> 292       2        7         muac             1         148.0
#> 293       3        7         muac             1         173.0
#> 294       4        7         muac             1         159.0
#> 295       5        7         muac             1         152.0
#> 296       6        7         muac             1         169.0
#> 297       7        7         muac             1         139.0
#> 298       8        7         muac             1         170.0
#> 299       9        7         muac             1         158.0
#> 300      10        7         muac             1         156.0
#> 301       1        8         muac             1         170.0
#> 302       2        8         muac             1         150.0
#> 303       3        8         muac             1         176.0
#> 304       4        8         muac             1         151.0
#> 305       5        8         muac             1         147.0
#> 306       6        8         muac             1         170.0
#> 307       7        8         muac             1         141.0
#> 308       8        8         muac             1         171.0
#> 309       9        8         muac             1         159.0
#> 310      10        8         muac             1         166.0
#> 311       1        9         muac             1         162.0
#> 312       2        9         muac             1         154.0
#> 313       3        9         muac             1         166.0
#> 314       4        9         muac             1         148.0
#> 315       5        9         muac             1         147.0
#> 316       6        9         muac             1         162.0
#> 317       7        9         muac             1         111.0
#> 318       8        9         muac             1         167.0
#> 319       9        9         muac             1         181.0
#> 320      10        9         muac             1         150.0
#> 321       1       10         muac             1         169.0
#> 322       2       10         muac             1         154.0
#> 323       3       10         muac             1         175.0
#> 324       4       10         muac             1         155.0
#> 325       5       10         muac             1         150.0
#> 326       6       10         muac             1         170.0
#> 327       7       10         muac             1         141.0
#> 328       8       10         muac             1         169.0
#> 329       9       10         muac             1         159.0
#> 330      10       10         muac             1         155.0
#> 331       1        0       weight             2          16.0
#> 332       2        0       weight             2          14.3
#> 333       3        0       weight             2          16.7
#> 334       4        0       weight             2          14.2
#> 335       5        0       weight             2          14.6
#> 336       6        0       weight             2          16.1
#> 337       7        0       weight             2          10.5
#> 338       8        0       weight             2          16.4
#> 339       9        0       weight             2          14.3
#> 340      10        0       weight             2          13.5
#> 341       1        1       weight             2          15.9
#> 342       2        1       weight             2          14.2
#> 343       3        1       weight             2          16.1
#> 344       4        1       weight             2          13.4
#> 345       5        1       weight             2          14.6
#> 346       6        1       weight             2          15.8
#> 347       7        1       weight             2          10.1
#> 348       8        1       weight             2          15.5
#> 349       9        1       weight             2          14.1
#> 350      10        1       weight             2          13.5
#> 351       1        2       weight             2          15.9
#> 352       2        2       weight             2          14.2
#> 353       3        2       weight             2          16.6
#> 354       4        2       weight             2          14.0
#> 355       5        2       weight             2          14.6
#> 356       6        2       weight             2          16.2
#> 357       7        2       weight             2          10.5
#> 358       8        2       weight             2          16.2
#> 359       9        2       weight             2          14.2
#> 360      10        2       weight             2          13.6
#> 361       1        3       weight             2          15.8
#> 362       2        3       weight             2          14.2
#> 363       3        3       weight             2          16.6
#> 364       4        3       weight             2          14.1
#> 365       5        3       weight             2          14.5
#> 366       6        3       weight             2          16.1
#> 367       7        3       weight             2          10.5
#> 368       8        3       weight             2          16.3
#> 369       9        3       weight             2          14.0
#> 370      10        3       weight             2          13.5
#> 371       1        4       weight             2          15.9
#> 372       2        4       weight             2          13.9
#> 373       3        4       weight             2          16.2
#> 374       4        4       weight             2          14.1
#> 375       5        4       weight             2          14.5
#> 376       6        4       weight             2          15.9
#> 377       7        4       weight             2          10.5
#> 378       8        4       weight             2          16.1
#> 379       9        4       weight             2          14.1
#> 380      10        4       weight             2          13.5
#> 381       1        5       weight             2          15.8
#> 382       2        5       weight             2          14.2
#> 383       3        5       weight             2          16.7
#> 384       4        5       weight             2          14.1
#> 385       5        5       weight             2          14.6
#> 386       6        5       weight             2          16.2
#> 387       7        5       weight             2          10.4
#> 388       8        5       weight             2          16.3
#> 389       9        5       weight             2          14.1
#> 390      10        5       weight             2          13.4
#> 391       1        6       weight             2          15.9
#> 392       2        6       weight             2          14.4
#> 393       3        6       weight             2          16.3
#> 394       4        6       weight             2          14.2
#> 395       5        6       weight             2          14.6
#> 396       6        6       weight             2          16.2
#> 397       7        6       weight             2          10.5
#> 398       8        6       weight             2          16.3
#> 399       9        6       weight             2          14.1
#> 400      10        6       weight             2          13.6
#> 401       1        7       weight             2          15.9
#> 402       2        7       weight             2          14.1
#> 403       3        7       weight             2          16.7
#> 404       4        7       weight             2          14.3
#> 405       5        7       weight             2          14.7
#> 406       6        7       weight             2          16.1
#> 407       7        7       weight             2          10.7
#> 408       8        7       weight             2          16.4
#> 409       9        7       weight             2          14.3
#> 410      10        7       weight             2          13.7
#> 411       1        8       weight             2          15.9
#> 412       2        8       weight             2          14.2
#> 413       3        8       weight             2          16.6
#> 414       4        8       weight             2          14.2
#> 415       5        8       weight             2          14.6
#> 416       6        8       weight             2          16.1
#> 417       7        8       weight             2          10.5
#> 418       8        8       weight             2          16.3
#> 419       9        8       weight             2          14.2
#> 420      10        8       weight             2          13.7
#> 421       1        9       weight             2          15.9
#> 422       2        9       weight             2          14.2
#> 423       3        9       weight             2          16.6
#> 424       4        9       weight             2          14.2
#> 425       5        9       weight             2          14.6
#> 426       6        9       weight             2          16.1
#> 427       7        9       weight             2          10.5
#> 428       8        9       weight             2          16.3
#> 429       9        9       weight             2          14.2
#> 430      10        9       weight             2          13.7
#> 431       1       10       weight             2          15.0
#> 432       2       10       weight             2          14.3
#> 433       3       10       weight             2          16.2
#> 434       4       10       weight             2          14.1
#> 435       5       10       weight             2          14.6
#> 436       6       10       weight             2          16.1
#> 437       7       10       weight             2          10.5
#> 438       8       10       weight             2          16.3
#> 439       9       10       weight             2          14.2
#> 440      10       10       weight             2          13.7
#> 441       1        0       height             2         102.3
#> 442       2        0       height             2          94.3
#> 443       3        0       height             2          97.2
#> 444       4        0       height             2          93.2
#> 445       5        0       height             2          99.1
#> 446       6        0       height             2          98.3
#> 447       7        0       height             2          83.5
#> 448       8        0       height             2          98.0
#> 449       9        0       height             2          92.9
#> 450      10        0       height             2          94.2
#> 451       1        1       height             2         101.8
#> 452       2        1       height             2          94.3
#> 453       3        1       height             2          96.8
#> 454       4        1       height             2          93.0
#> 455       5        1       height             2          99.4
#> 456       6        1       height             2          19.2
#> 457       7        1       height             2          14.1
#> 458       8        1       height             2          97.8
#> 459       9        1       height             2          94.4
#> 460      10        1       height             2          93.9
#> 461       1        2       height             2         101.1
#> 462       2        2       height             2          93.6
#> 463       3        2       height             2          95.9
#> 464       4        2       height             2          92.8
#> 465       5        2       height             2          99.3
#> 466       6        2       height             2          98.2
#> 467       7        2       height             2         134.0
#> 468       8        2       height             2          97.2
#> 469       9        2       height             2          93.8
#> 470      10        2       height             2          94.2
#> 471       1        3       height             2         102.5
#> 472       2        3       height             2          94.3
#> 473       3        3       height             2          97.1
#> 474       4        3       height             2          93.1
#> 475       5        3       height             2         100.0
#> 476       6        3       height             2          98.6
#> 477       7        3       height             2          84.1
#> 478       8        3       height             2          98.2
#> 479       9        3       height             2          94.2
#> 480      10        3       height             2          95.2
#> 481       1        4       height             2         102.2
#> 482       2        4       height             2          94.8
#> 483       3        4       height             2          97.1
#> 484       4        4       height             2          93.1
#> 485       5        4       height             2          99.3
#> 486       6        4       height             2          98.4
#> 487       7        4       height             2          83.3
#> 488       8        4       height             2          97.8
#> 489       9        4       height             2          94.2
#> 490      10        4       height             2          93.7
#> 491       1        5       height             2         101.2
#> 492       2        5       height             2          94.1
#> 493       3        5       height             2          95.1
#> 494       4        5       height             2          92.8
#> 495       5        5       height             2          99.2
#> 496       6        5       height             2          97.2
#> 497       7        5       height             2          83.3
#> 498       8        5       height             2          97.6
#> 499       9        5       height             2          93.2
#> 500      10        5       height             2          93.7
#> 501       1        6       height             2         101.5
#> 502       2        6       height             2          93.1
#> 503       3        6       height             2          96.5
#> 504       4        6       height             2          91.7
#> 505       5        6       height             2          99.8
#> 506       6        6       height             2          97.7
#> 507       7        6       height             2          83.9
#> 508       8        6       height             2          97.1
#> 509       9        6       height             2          92.0
#> 510      10        6       height             2          94.1
#> 511       1        7       height             2         101.6
#> 512       2        7       height             2          94.2
#> 513       3        7       height             2          96.9
#> 514       4        7       height             2          93.0
#> 515       5        7       height             2          99.5
#> 516       6        7       height             2          98.0
#> 517       7        7       height             2          83.4
#> 518       8        7       height             2          97.7
#> 519       9        7       height             2          93.7
#> 520      10        7       height             2          93.5
#> 521       1        8       height             2         102.3
#> 522       2        8       height             2          94.3
#> 523       3        8       height             2          97.0
#> 524       4        8       height             2          92.5
#> 525       5        8       height             2         101.0
#> 526       6        8       height             2          98.4
#> 527       7        8       height             2          84.5
#> 528       8        8       height             2          98.0
#> 529       9        8       height             2          93.9
#> 530      10        8       height             2          94.4
#> 531       1        9       height             2         102.4
#> 532       2        9       height             2          93.7
#> 533       3        9       height             2          97.1
#> 534       4        9       height             2          92.6
#> 535       5        9       height             2          99.4
#> 536       6        9       height             2          98.5
#> 537       7        9       height             2          94.1
#> 538       8        9       height             2          97.2
#> 539       9        9       height             2          93.6
#> 540      10        9       height             2          93.6
#> 541       1       10       height             2         102.0
#> 542       2       10       height             2          93.9
#> 543       3       10       height             2          97.1
#> 544       4       10       height             2          93.0
#> 545       5       10       height             2          98.6
#> 546       6       10       height             2          97.8
#> 547       7       10       height             2          84.5
#> 548       8       10       height             2          97.5
#> 549       9       10       height             2          94.4
#> 550      10       10       height             2          94.6
#> 551       1        0         muac             2         170.0
#> 552       2        0         muac             2         158.0
#> 553       3        0         muac             2         173.0
#> 554       4        0         muac             2         164.0
#> 555       5        0         muac             2         160.0
#> 556       6        0         muac             2         170.0
#> 557       7        0         muac             2         145.0
#> 558       8        0         muac             2         171.0
#> 559       9        0         muac             2         160.0
#> 560      10        0         muac             2         160.0
#> 561       1        1         muac             2         163.0
#> 562       2        1         muac             2         154.0
#> 563       3        1         muac             2         176.0
#> 564       4        1         muac             2         154.0
#> 565       5        1         muac             2         154.0
#> 566       6        1         muac             2         169.0
#> 567       7        1         muac             2         140.0
#> 568       8        1         muac             2         169.0
#> 569       9        1         muac             2         161.0
#> 570      10        1         muac             2         154.0
#> 571       1        2         muac             2         164.0
#> 572       2        2         muac             2         143.0
#> 573       3        2         muac             2         174.0
#> 574       4        2         muac             2         151.0
#> 575       5        2         muac             2         149.0
#> 576       6        2         muac             2         163.0
#> 577       7        2         muac             2         141.0
#> 578       8        2         muac             2         165.0
#> 579       9        2         muac             2         150.0
#> 580      10        2         muac             2         152.0
#> 581       1        3         muac             2         174.0
#> 582       2        3         muac             2         155.0
#> 583       3        3         muac             2         184.0
#> 584       4        3         muac             2         154.0
#> 585       5        3         muac             2          15.0
#> 586       6        3         muac             2         170.0
#> 587       7        3         muac             2         146.0
#> 588       8        3         muac             2         170.0
#> 589       9        3         muac             2         153.0
#> 590      10        3         muac             2         154.0
#> 591       1        4         muac             2         170.0
#> 592       2        4         muac             2         150.0
#> 593       3        4         muac             2         180.0
#> 594       4        4         muac             2         157.0
#> 595       5        4         muac             2         153.0
#> 596       6        4         muac             2         167.0
#> 597       7        4         muac             2         141.0
#> 598       8        4         muac             2         171.0
#> 599       9        4         muac             2         164.0
#> 600      10        4         muac             2         155.0
#> 601       1        5         muac             2         172.0
#> 602       2        5         muac             2         157.0
#> 603       3        5         muac             2         177.0
#> 604       4        5         muac             2         155.0
#> 605       5        5         muac             2         150.0
#> 606       6        5         muac             2         167.0
#> 607       7        5         muac             2         143.0
#> 608       8        5         muac             2         173.0
#> 609       9        5         muac             2         157.0
#> 610      10        5         muac             2         196.0
#> 611       1        6         muac             2         173.0
#> 612       2        6         muac             2         151.0
#> 613       3        6         muac             2         167.0
#> 614       4        6         muac             2         155.0
#> 615       5        6         muac             2         155.0
#> 616       6        6         muac             2         175.0
#> 617       7        6         muac             2         134.0
#> 618       8        6         muac             2         176.0
#> 619       9        6         muac             2         153.0
#> 620      10        6         muac             2         152.0
#> 621       1        7         muac             2         160.0
#> 622       2        7         muac             2         155.0
#> 623       3        7         muac             2         167.0
#> 624       4        7         muac             2         151.0
#> 625       5        7         muac             2         140.0
#> 626       6        7         muac             2         164.0
#> 627       7        7         muac             2         139.0
#> 628       8        7         muac             2         175.0
#> 629       9        7         muac             2         186.0
#> 630      10        7         muac             2         152.0
#> 631       1        8         muac             2         169.0
#> 632       2        8         muac             2         152.0
#> 633       3        8         muac             2         173.0
#> 634       4        8         muac             2         159.0
#> 635       5        8         muac             2         149.0
#> 636       6        8         muac             2         165.0
#> 637       7        8         muac             2         141.0
#> 638       8        8         muac             2         172.0
#> 639       9        8         muac             2         158.0
#> 640      10        8         muac             2         156.0
#> 641       1        9         muac             2         162.0
#> 642       2        9         muac             2         159.0
#> 643       3        9         muac             2         166.0
#> 644       4        9         muac             2         148.0
#> 645       5        9         muac             2         147.0
#> 646       6        9         muac             2         162.0
#> 647       7        9         muac             2         112.0
#> 648       8        9         muac             2         176.0
#> 649       9        9         muac             2         181.0
#> 650      10        9         muac             2         150.0
#> 651       1       10         muac             2         165.0
#> 652       2       10         muac             2         152.0
#> 653       3       10         muac             2         171.0
#> 654       4       10         muac             2         153.0
#> 655       5       10         muac             2         147.0
#> 656       6       10         muac             2         165.0
#> 657       7       10         muac             2         141.0
#> 658       8       10         muac             2         171.0
#> 659       9       10         muac             2         161.0
#> 660      10       10         muac             2         156.0