Standardisation test data - sample 3

Example dataset from the Standardized Monitoring and Assessment of Relief and Transitions (SMART) capacity building toolbox found at <http://smartmethodology.org/survey-planning-tools/smart-capacity-building-toolbox/>. This dataset is the same as the smartStd dataset but has been kept in the original wide format.

Usage

smartWide

Format

A data frame with 67 columns and 10 rows

Source

Standardized Monitoring and Assessment of Relief and Transitions (SMART) capacity building toolbox found at <http://smartmethodology.org/survey-planning-tools/smart-capacity-building-toolbox/>.

Examples

smartWide
#> # A tibble: 10 × 67
#>    subject weight_1_supervisor weight_2_supervisor height_1_supervisor
#>      <int>               <dbl>               <dbl>               <dbl>
#>  1       1                15.9                16                 102  
#>  2       2                14.2                14.3                94.8
#>  3       3                16.7                16.7                97.2
#>  4       4                14.1                14.2                93.1
#>  5       5                14.8                14.6                99.6
#>  6       6                16.2                16.1                98.9
#>  7       7                10.5                10.5                83.5
#>  8       8                16.5                16.4                98  
#>  9       9                14.1                14.3                95.1
#> 10      10                13.5                13.5                94.7
#> # ℹ 63 more variables: height_2_supervisor <dbl>, muac_1_supervisor <dbl>,
#> #   muac_2_supervisor <dbl>, weight_1_enumerator1 <dbl>,
#> #   weight_2_enumerator1 <dbl>, height_1_enumerator1 <dbl>,
#> #   height_2_enumerator1 <dbl>, muac_1_enumerator1 <dbl>,
#> #   muac_2_enumerator1 <dbl>, weight_1_enumerator2 <dbl>,
#> #   weight_2_enumerator2 <dbl>, height_1_enumerator2 <dbl>,
#> #   height_2_enumerator2 <dbl>, muac_1_enumerator2 <dbl>, …