stature

Example dataset from Ulijaszek and Kerr (1999) containing repeat measurements of stature m carried out by four observers on ten subjects.

stature

Format

A data frame with 8 columns and 10 rows:

subject

Subject unique identification number

m1

Stature measurements of first observer

m2

Stature measurements of second observer

m3

Stature measurements of third observer

m4

Stature measurements of fourth observer

f1

Results of the first function in the group TEM equation

f2

Results of the second function in the group TEM equation

diff

Difference between f2 and f2 (f2 - f1)

Source

1. Ulijaszek SJ, Kerr DA. Anthropometric measurement error and the assessment of nutritional status. Br J Nutr. 2007;82: 165–13. doi:10.1017/S0007114599001348