Calculate relative TEM of measurements by multiple observers.

Calculate relative TEM of measurements by multiple observers.

Usage

calculate_relative_tem(tem, mean_value)

Arguments

tem

A numeric vector or data frame of technical error of measurements produced from applying [calculate_tem_cohort()].

mean_value

A numeric vector or data frame of mean of measurements produced from applying [calculate_mean()].

Value

A vector or data frame of calculated relative TEM per observer and per measurement type.

Examples

## Calculate relative TEM for weight using the smartStd dataset
  tem <- calculate_tem_cohort(df = smartStd,
                              m1 = "weight1",
                              m2 = "weight2",
                              index = "observer",
                              n = 10)

  x <- smartStdLong[smartStdLong$measure_type == "weight", ]

  mean_value <- calculate_mean(measures = x$measure_value,
                               index = x[ , "observer"])

  rel_tem <- calculate_relative_tem(tem = tem, mean_value = mean_value)
#> Warning: ‘/’ not meaningful for factors
  rel_tem
#>    observer       tem
#> 1        NA 0.5501370
#> 2        NA 2.3430872
#> 3        NA 0.8126483
#> 4        NA 0.4337829
#> 5        NA 1.1606699
#> 6        NA 9.7163749
#> 7        NA 0.9408907
#> 8        NA 0.6987351
#> 9        NA 0.7782767
#> 10       NA 0.7477441
#> 11       NA 2.0123765